feat: 데이터 타입 수정

+++ src/main/java/toy/board/blog/dto/request/CreateCategoryRequest.java
@@ -3,7 +3,7 @@ package toy.board.blog.dto.request;
 import jakarta.validation.constraints.NotEmpty;
 import toy.board.entity.Category;
 
-public record CreateCategoryRequest(Long parentId,
+public record CreateCategoryRequest(Integer parentId,
                                     @NotEmpty(message = "카테고리의 이름은 필수입니다.") String name
 ) {
     public Category toCategory(Category parentCategory) {
